Fix the model structure
Fixes #6 .
To start a new model use
    `python manage.py startapp <name-of-app>`
Then start edit in model.py inside the new folder of <name-of-app>.For a
model write all field in the same class. See the changes of this commit
to know how to write a model. Use the exact name for the field rather
than _text suffix.