Lay out foundation for templates
Basic foundation for templates is setup. Similarly the templates for
account activities are set to inherit global base.
The dynamic content fetching is enabled for the django homepage.
The relationship score is used to show most related data to the user.

After signup of user,they are requested to select category and tags.
The category auto fill ajax fetch is set up

Run pip install to update project.

It is recommended to drop the data base and migrate again.